Add python_requires to prevent people installing on Py2 (#12)
When people spot a plugin they like, they go to the plugin manager to install it.
But it doesn't show up on Py2 installs now due to the compatibility string. So they install it manually.

This saves you from the headaches of 'Installed but not loading' since it refuses, with an error about Python compatibility.